@@ -7,12 +7,12 @@
     "run-ios": "yarn react-native run-ios",
     "run-android": "yarn react-native run-android",
     "info": "yarn react-native info",
-    "test": "jest",
+    "test": "jest --detectOpenHandles",
     "linter": "yarn eslint ./src",
-    "upgrade-deps": "yarn upgrade-interactive --latest"
+    "upgrade-deps": "yarn upgrade-interactive --latest",
+    "mobx": "mobx-devtools"
   },
   "dependencies": {
-    "global": "^4.3.2",
     "lottie-react-native": "^2.6.1",
     "mobx": "^5.9.0",
     "mobx-devtools": "^0.9.21",
@@ -24,9 +24,10 @@
     "moment-duration-format": "2.2.2",
     "prop-types": "^15.7.2",
     "react": "16.8.4",
-    "react-native": "0.59.0",
+    "react-native": "0.59.10",
     "react-native-circular-progress": "^1.1.0",
     "react-native-elements": "^1.1.0",
+    "react-native-firebase": "^5.5.4",
     "react-native-gesture-handler": "^1.1.0",
     "react-native-linear-gradient": "^2.5.4",
     "react-native-sound": "^0.10.12",
@@ -36,21 +37,28 @@
     "react-timer-machine": "^1.0.2"
   },
   "devDependencies": {
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "24.5.0",
-    "eslint": "^5.16.0",
+    "@react-native-community/eslint-config": "^0.0.5",
+    "@types/react": "^16.8.24",
+    "@types/react-native": "^0.60.3",
+    "@types/react-navigation": "^3.0.7",
+    "@babel/core": "^7.4.5",
+    "@babel/runtime": "^7.4.5",
+    "babel-jest": "^24.8.0",
+    "eslint": "^6.0.1",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-prettier": "^4.3.0",
     "eslint-plugin-import": "^2.17.3",
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-react": "^7.13.0",
+    "eslint-plugin-react-hooks": "^1.6.1",
     "eslint-plugin-react-native": "^3.7.0",
-    "jest": "24.5.0",
-    "metro-react-native-babel-preset": "0.53.0",
+    "jest": "^24.8.0",
+    "metro-react-native-babel-preset": "^0.54.1",
     "prettier": "^1.17.1",
     "prettier-eslint": "^8.8.2",
+    "react-dom": "16.8.4",
+    "react-native-testing-library": "^1.11.1",
     "react-test-renderer": "16.8.4"
   },
   "jest": {
