@@ -1,61 +1,50 @@
 import React from 'react';
 import { StyleSheet, Text } from 'react-native';
 import PropTypes from 'prop-types';
+import Sound from 'react-native-sound';
 import { ProgressContainer, TimerMachineContainer } from '../containers';
-
-var Sound = require('react-native-sound');
+import { fontSizes, fontFamilies, colors } from '../../../utils/theme';
 
 const styles = StyleSheet.create({
   titleText: {
-    fontSize: 100,
+    fontSize: fontSizes.xxxLarge,
     position: 'absolute',
-    fontFamily: 'League Gothic',
+    fontFamily: fontFamilies.leagueGothic,
     fontWeight: 'bold',
-    color: 'white',
+    color: colors.secondaryDark,
     alignSelf: 'center',
-  },
-  container: {
-    position: 'absolute',
-    alignSelf: 'center',
-    height: 20,
-    justifyContent: 'center',
-  },
-  secondContainer: {
-    position: 'absolute',
-    alignSelf: 'center',
-    height: 20,
-    justifyContent: 'center',
   },
 });
 
 const Countdown = props => {
-  const { count, isRest, fill, fillComplete, _onFinish } = props;
-  const backgroundColors = isRest ? '#38ef7d' : '#6DD5FA';
-  const backgroundColorsOff = isRest ? '#1D4350' : '#1D4350';
+  const { fill, fillComplete, _onFinish, currentCircleColor } = props;
+  // const backgroundColors = isRest ? colors.restGreen : colors.restBlue;
+  // const backgroundColorsOff = isRest ? colors.restGreenDeep : colors.secondaryLight;
 
   Sound.setCategory('Playback');
 
   return (
     <React.Fragment>
       <Text style={styles.titleText}>
-        <TimerMachineContainer count={count} _onFinish={_onFinish} />
+        <TimerMachineContainer _onFinish={_onFinish} />
       </Text>
       <ProgressContainer
         fill={fill}
         fillComplete={fillComplete}
-        tintColor={backgroundColors}
-        backgroundColor={backgroundColorsOff}
+        tintColor={currentCircleColor}
+        // backgroundColor={backgroundColorsOff}
       />
     </React.Fragment>
   );
 };
 
 Countdown.propTypes = {
-  count: PropTypes.number.isRequired,
-  isRest: PropTypes.bool.isRequired,
+  // count: PropTypes.number.isRequired,
+  // isRest: PropTypes.bool.isRequired,
   fill: PropTypes.number.isRequired,
   fillComplete: PropTypes.number.isRequired,
   _onFinish: PropTypes.func.isRequired,
+  currentCircleColor: PropTypes.string.isRequired,
 };
 
 Countdown.defaultProps = {};
