@@ -1,13 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { View } from 'react-native';
-
 import { observer, inject } from 'mobx-react';
-import moment from 'moment';
-import momentDurationFormatSetup from 'moment-duration-format';
-import { Congrats } from '../components';
 
-momentDurationFormatSetup(moment);
+import { Congrats } from '../components';
 
 class CongratsContainer extends React.Component {
   render() {
@@ -15,16 +10,14 @@ class CongratsContainer extends React.Component {
     const { training, rest, cycle, serie, endTime } = session;
 
     return (
-      <View>
-        <Congrats
-          training={training}
-          rest={rest}
-          cycle={cycle}
-          serie={serie}
-          endTime={endTime}
-          onPress={onPress}
-        />
-      </View>
+      <Congrats
+        training={training}
+        rest={rest}
+        cycle={cycle}
+        serie={serie}
+        endTime={endTime}
+        onPress={onPress}
+      />
     );
   }
 }
