@@ -5,25 +5,19 @@ const styles = StyleSheet.create({
     flex: 1,
   },
   top: {
-    flex: 1,
+    flex: 0.2,
     justifyContent: 'flex-end',
     flexDirection: 'row',
-    // borderColor: 'blue',
-    // borderWidth: 3,
   },
   middle: {
     flex: 7,
     flexDirection: 'row',
     justifyContent: 'center',
-    // borderColor: 'red',
-    // borderWidth: 3,
   },
   bottom: {
-    flex: 4,
+    flex: 4.8,
     flexDirection: 'row',
     justifyContent: 'space-around',
-    // borderColor: 'red',
-    // borderWidth: 3,
   },
   linearGradient: {
     flex: 1,
