@@ -2,8 +2,8 @@ import React from 'react';
 import { View, StyleSheet } from 'react-native';
 import PropTypes from 'prop-types';
 
-import moment from 'moment';
-import momentDurationFormatSetup from 'moment-duration-format';
+// import moment from 'moment';
+// import momentDurationFormatSetup from 'moment-duration-format';
 import {
   CountdownContainer,
   LapCycleContainer,
@@ -11,7 +11,7 @@ import {
   TimeElapsedContainer,
 } from '../containers';
 
-momentDurationFormatSetup(moment);
+// momentDurationFormatSetup(moment);
 
 const styles = StyleSheet.create({
   container: {
@@ -32,24 +32,24 @@ const styles = StyleSheet.create({
 });
 
 const Timer = props => {
-  const { inProgress, goToCongratsFromHome } = props;
+  const { goToCongratsFromHome } = props;
   return (
     <View style={styles.container}>
       <View style={styles.middleTop}>
-        <CountdownContainer inProgress={inProgress} goToCongratsFromHome={goToCongratsFromHome} />
+        <CountdownContainer goToCongratsFromHome={goToCongratsFromHome} />
       </View>
       <View style={styles.middleMiddle}>
         <LapSerieContainer />
         <LapCycleContainer />
       </View>
       <View style={styles.middleBottom}>
-        <TimeElapsedContainer inProgress={inProgress} goToCongratsFromHome={goToCongratsFromHome} />
+        <TimeElapsedContainer />
       </View>
     </View>
   );
 };
 Timer.propTypes = {
-  inProgress: PropTypes.bool.isRequired,
+  // inProgress: PropTypes.bool.isRequired,
   goToCongratsFromHome: PropTypes.func.isRequired,
 };
 
