@@ -2,10 +2,11 @@ import React from 'react';
 import { View, StyleSheet } from 'react-native';
 import PropTypes from 'prop-types';
 import Timer from './Timer';
+import { fontSizes } from '../../../utils/theme';
 
 const styles = StyleSheet.create({
   text: {
-    fontSize: 40,
+    fontSize: fontSizes.xLarge,
   },
   middleLeft: {
     flex: 2,
@@ -19,19 +20,19 @@ const styles = StyleSheet.create({
 });
 
 const WorkoutSpace = props => {
-  const { inProgress, goToCongratsFromHome } = props;
+  const { goToCongratsFromHome } = props;
   return (
     <>
       <View style={styles.middleLeft} />
       <View style={styles.middleCenter}>
-        <Timer inProgress={inProgress} goToCongratsFromHome={goToCongratsFromHome} />
+        <Timer goToCongratsFromHome={goToCongratsFromHome} />
       </View>
       <View style={styles.middleRight} />
     </>
   );
 };
 WorkoutSpace.propTypes = {
-  inProgress: PropTypes.bool.isRequired,
+  // inProgress: PropTypes.bool.isRequired,
   goToCongratsFromHome: PropTypes.func.isRequired,
 };
 
