@@ -1,56 +1,45 @@
 import React from 'react';
 import { observer, inject } from 'mobx-react';
 import PropTypes from 'prop-types';
+
+import { colors } from 'HiitTimer/src/utils/theme';
 import { Countdown } from '../components';
 
 class CountdownContainer extends React.Component {
-  _onFinish = () => {
+  onFinish = () => {
     const {
       session: { timeCompleteWorkout, timePased, setStop },
       goToCongratsFromHome,
+      // Reset progress
     } = this.props;
-    if (timeCompleteWorkout == timePased) {
+    if (timeCompleteWorkout === timePased) {
       setStop();
       goToCongratsFromHome();
     }
   };
+
   render() {
     const { session } = this.props;
     const {
-      training,
-      rest,
       isRest,
-      isPlay,
-      isStop,
-      increaseSerie,
-      currentTime,
-      resetTime,
-      setStop,
-      timePased,
-      inProgress,
-      timeCompleteWorkout,
-      playSound,
+      timePasedWorkout,
+      timePasedPerSerie,
+      // playSound,
     } = session;
 
-    const count = isRest ? rest : training;
-    const fill = -((currentTime / count) * 100 - 100);
-    const fillComplete = (timePased / timeCompleteWorkout) * 100;
+    const currentCircleColor = isRest ? colors.restGreen : colors.restBlue;
+    // const count = isRest ? rest : training;
+    // console.log({ currentTime, count });
+
+    // console.log({ timePasedPerSerie, timePasedWorkout });
 
     return (
       <Countdown
-        count={count}
-        isRest={isRest}
-        isPlay={isPlay}
-        isStop={isStop}
-        increaseSerie={increaseSerie}
-        currentTime={currentTime}
-        resetTime={resetTime}
-        setStop={setStop}
-        inProgress={inProgress}
-        fill={fill}
-        fillComplete={fillComplete}
-        _onFinish={this._onFinish}
-        playSound={playSound}
+        currentCircleColor={currentCircleColor}
+        fill={timePasedPerSerie}
+        fillComplete={timePasedWorkout}
+        _onFinish={this.onFinish}
+        // playSound={playSound}
       />
     );
   }
