@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import { PlayPause } from '../components';
 
 class PlayPauseContainer extends React.Component {
-  _onPlay = () => {
+  onPlay = () => {
     const {
       session: { setPlay, inProgress },
       onPress,
@@ -22,7 +22,7 @@ class PlayPauseContainer extends React.Component {
     return (
       <PlayPause
         isPlay={session.isPlay}
-        onPlay={this._onPlay}
+        onPlay={this.onPlay}
         onPause={session.setPause}
         onStop={session.setStop}
         setInProgress={session.setInProgress}
