@@ -7,8 +7,8 @@ import { TimeElapsed } from '../components';
 class TimeElapsedContainer extends React.Component {
   render() {
     const { session } = this.props;
-    const { isPlay, isStop, increaseTime } = session;
-    return <TimeElapsed isPlay={isPlay} isStop={isStop} increaseTime={increaseTime} />;
+    const { isPlay, isStop } = session;
+    return <TimeElapsed isPlay={isPlay} isStop={isStop} />;
   }
 }
 
