@@ -1,8 +1,10 @@
+/* eslint import/no-cycle: [2, { maxDepth: 1 }] */
+
 export { default as CountdownContainer } from './CountdownContainer';
 export { default as LapCycleContainer } from './LapCycleContainer';
 export { default as LapSerieContainer } from './LapSerieContainer';
 export { default as PlayPauseContainer } from './PlayPauseContainer';
 export { default as TimeElapsedContainer } from './TimeElapsedContainer';
-export { default as CircularProgressContainer } from './CircularProgressContainer';
 export { default as ProgressContainer } from './ProgressContainer';
+export { default as CircularProgressContainer } from './CircularProgressContainer';
 export { default as TimerMachineContainer } from './TimerMachineContainer';
