@@ -1,20 +1,20 @@
 import { StyleSheet } from 'react-native';
+// import { space } from '../../utils/theme';
+// import { space } from 'HiitTimer/src/utils/theme';
+// import { space } from 'src/utils/theme';
 
 const styles = StyleSheet.create({
-  top: {
-    flexDirection: 'row',
-    paddingLeft: 8,
-    paddingRight: 8,
-  },
-  middle: {},
-  bottom: {
-    marginTop: 10,
-    paddingLeft: 8,
-    paddingRight: 8,
-  },
-  linearGradient: {
+  container: {
     flex: 1,
   },
+  top: {
+    flex: 0.1,
+    justifyContent: 'center',
+  },
+  middle: {
+    flex: 0.8,
+  },
+  bottom: {},
 });
 
 export default styles;
